+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - OS environment expansion in ${VAR}
 - Global options:
   - `use_os_env_for_expand` to disable OS environment expansion
+  - `environment`.`extra_os` to define environment variables
 
 ### Changed
 - KiCad v6/7 schematic: